% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bces.R
\name{bces}
\alias{bces}
\title{Causal Mediation with BCES estimator}
\usage{
bces(
  gamma_hat,
  tau_hat,
  sd_u,
  sd_v,
  confid = 0.05,
  bootstrap = TRUE,
  b = 1000,
  seed = 320
)
}
\arguments{
\item{gamma_hat}{a vector of the effect of the treatment on the mediator.}

\item{tau_hat}{a vector of the total effect of the treatment on the outcome.}

\item{sd_u}{a vector of the standard error of gamma, the effect of the treatment on the mediator.}

\item{sd_v}{a vector of the standard error of tau, the total effect of the treatment on the outcome.}

\item{confid}{the size of the test. The default is 0.05.}

\item{bootstrap}{logical value. Whether to use bootstrap.}

\item{b}{the number of bootstrap replicates. The default number is 1000.}

\item{seed}{a single value passed to \code{set.seed}.}
}
\value{
beta estimated value of beta.

alpha estimated value of alpha.

sd_alpha standard error of the estimated alpha.

sd_beta standard error of the estimated beta.

ci_alpha confidence Interval of alpha based on analytic formula.

ci_beta confidence Interval of beta based on analytic formula.

boot_ci_alpha confidence Interval of alpha based on bootstrap.

boot_ci_beta confidence Interval of beta based on bootstrap.

p_alpha P-value of alpha based on analytic formula.

p_beta P-value of beta based on analytic formula.

boot_p_alpha P-value of alpha based on bootstrap.

boot_p_beta P-value of beta based on bootstrap.

ave_med the estimated average mediation effect.
}
\description{
This function estimates the the beta (key part of the causal mediation effect) with the BCES estimator following Fu (2023).In general,
}
\examples{
\dontrun{
data(example_dat)  # input data
example_dat <- as.data.frame(example_dat)
bces(example_dat$gamma_hat,example_dat$tau_hat,example_dat$sd_u,example_dat$sd_v) # simex estimator

tmp <- bces(example_dat$gamma_hat,example_dat$tau_hat,sd_u) # save the result
tmp$beta # extract the value
}

}
