% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_Data.R
\docType{data}
\name{example_dat}
\alias{example_dat}
\title{Example dataset}
\format{
A data frame with 100 observations and 4 variables:
\describe{
\item{gamma_hat}{the effect of the treatment on the mediato}
\item{tau_hat}{the total effect of the treatment on the outcome}
\item{sd_u}{the standard error of gamma, the effect of the treatment on the mediator}
\item{sd_v}{the standard error of tau, the total effect of the treatment on the outcome}
}
}
\usage{
example_dat
}
\description{
A dataset containing variables...
}
\keyword{datasets}
