% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mech.R
\name{plot_mech}
\alias{plot_mech}
\title{Plot Function for Causal Mediation}
\usage{
plot_mech(
  gamma_hat,
  tau_hat,
  sd_u,
  sd_v,
  confid = 0.05,
  b = 1000,
  fit = "SIMEX",
  seed = 320,
  legend = "TRUE",
  xlab = "ATE on the Mediator",
  ylab = "ATE on the Outcome",
  col_error_bar = "#219ebc",
  col_point = "#219ebc",
  col_line = c("#8ac926", "#fe7f2d"),
  size_title = 14,
  size_text = 12,
  size_legend = 12
)
}
\arguments{
\item{gamma_hat}{a vector of the treatment effect on the mediator.}

\item{tau_hat}{a vector of the total treatment effect on the outcome.}

\item{sd_u}{a vector of the standard error of \code{gamma_hat}, the treatment effect on the mediator.}

\item{sd_v}{a vector of the standard error of \code{tau_hat}, the total treatment effect on the outcome.}

\item{confid}{the size of the test. The default is 0.05.}

\item{b}{the number of bootstrap replicates. The default number is 1000.}

\item{fit}{the estimator to use. Default is "SIMEX"; additional options are "BCES" or "both". However, we suggest using "SIMEX".}

\item{seed}{a single value passed to \code{set.seed}.}

\item{legend}{logical value; whether to include legend in the figure. Default is "TRUE".}

\item{xlab}{the label on the x-axis. Default is "ATE on the Mediator".}

\item{ylab}{the label on the y-axis. Default is "ATE on the Outcome".}

\item{col_error_bar}{the color of the error bar. Default isn "#219ebc".}

\item{col_point}{the color of the point. Default isn "#219ebc".}

\item{col_line}{the color of the fitted line. Default is "#8ac926" and "#fe7f2d".}

\item{size_title}{the size of the title. Default is 14.}

\item{size_text}{the size of the text on the axis. Default is 12.}

\item{size_legend}{the size of the legend. Default is 12.}
}
\value{
Figure.
}
\description{
This function is designed to plot the figure of average mediation effects as outlined in Fu (2023). In general, the method transforms the challenging mediation problem into a simple linear regression problem without compromising the non-parametric nature. The resulting plot will display the observations of average treatment effects on both the mediator and the outcome, their corresponding standard errors, and the estimated regression line. The slope of this regression line represents the key parameter \code{beta}, which indicates the rate of change in the pure indirect effect relative to a one-unit change in the average treatment effects on the mediator. This parameter is crucial for constructing the average mediation effects.
}
\examples{
data(example_dat)  # input data
example_dat <- as.data.frame(example_dat)
plot_mech(example_dat$gamma_hat,example_dat$tau_hat,example_dat$sd_u,example_dat$sd_v)
}
