% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simest.R
\name{simest}
\alias{simest}
\title{Causal Mediation with SIMEX estimator}
\usage{
simest(gamma_hat, tau_hat, sd_u, b = 1000)
}
\arguments{
\item{gamma_hat}{a vector of the effect of the treatment on the mediator.}

\item{tau_hat}{a vector of the total effect of the treatment on the outcome.}

\item{sd_u}{a vector of the standard error of gamma, the effect of the treatment on the mediator.}

\item{b}{the number of bootstrap replicates. The default number is 1000.}
}
\value{
\code{ave_med} the estimated average mediation effects.

\code{beta} estimated value of beta.

\code{alpha} estimated value of alpha.

\code{se_alpha} standard error of the estimated alpha.

\code{se_beta} standard error of the estimated beta.

\code{p_alpha} P-value of alpha.

\code{p_beta} P-value of beta.
}
\description{
This function estimates the average mediation effects with the SIMEX estimator, following the methodology outlined in Fu (2023). In general, the method transforms the challenging mediation problem into a simple linear regression problem without compromising the non-parametric nature. To implement the SIMEX estimator, we assume that researchers have obtained multiple estimated average treatment effects on the mediator and outcome, and corresponding standard errors. To get those average treatment effects, one can use causal trees or meta-analysis, as described by Fu (2023). With those estimates as inputs, this function returns average mediation effects.
}
\examples{
data(example_dat)  # input data
example_dat <- as.data.frame(example_dat)
simest(example_dat$gamma_hat,example_dat$tau_hat,example_dat$sd_u) # simex estimator

# assign the result to the tmp
tmp <- simest(example_dat$gamma_hat,example_dat$tau_hat,example_dat$sd_u)
tmp$ave_med  # extract the value

}
\references{
Jiawei Fu. 2013. "Extract Mechanisms from Heterogeneous Effects: A New Identification Strategy for Mediation Analysis" \emph{Working Paper}.
}
