% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mechte_bces}
\alias{summary.mechte_bces}
\title{Summary Methods for mechte Results}
\usage{
\method{summary}{mechte_bces}(object, ...)
}
\arguments{
\item{object}{A \code{mechte_bces} or \code{mechte_simest} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The input object, invisibly.
}
\description{
Summary Methods for mechte Results
}
