% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mech.R
\name{plot_mech}
\alias{plot_mech}
\title{Plot the estimates by}
\usage{
plot_mech(
  gamma_hat,
  tau_hat,
  sd_u,
  sd_v,
  confid = 0.05,
  b = 1000,
  fit = "SIMEX",
  seed = 320,
  legend = "TRUE",
  xlab = "ATE on the Mediator",
  ylab = "ATE on the Outcome",
  col_error_bar = "#219ebc",
  col_point = "#219ebc",
  col_line = c("#8ac926", "#fe7f2d"),
  size_title = 14,
  size_text = 12,
  size_legend = 12
)
}
\arguments{
\item{gamma_hat}{a vector of the effect of the treatment on the mediator.}

\item{tau_hat}{a vector of the total effect of the treatment on the outcome.}

\item{sd_u}{a vector of the standard error of gamma, the effect of the treatment on the mediator.}

\item{sd_v}{a vector of the standard error of tau, the total effect of the treatment on the outcome.}

\item{confid}{the size of the test. The default is 0.05.}

\item{b}{the number of bootstrap replicates. The default number is 1000.}

\item{fit}{the estimator to use. Default is "SIMEX"; additional options are "BCES" or "both".}

\item{seed}{a single value passed to \code{set.seed}.}

\item{legend}{logical value; whether to include legend in the figure. Default is "TRUE".}

\item{xlab}{the label on the x-axis. Default is "ATE on the Mediator".}

\item{ylab}{the label on the y-axis. Default is "ATE on the Outcome".}

\item{col_error_bar}{the color of the error bar. Default isn "#219ebc".}

\item{col_point}{the color of the point. Default isn "#219ebc".}

\item{col_line}{the color of the fitted line. Default is "#8ac926" and "#fe7f2d".}

\item{size_title}{the size of the title. Default is 14.}

\item{size_text}{the size of the text on the axis. Default is 12.}

\item{size_legend}{the size of the legend. Default is 12.}
}
\value{
Figure.
}
\description{
Plot the estimates by
}
\examples{
\dontrun{
data(example_dat)  # input data
example_dat <- as.data.frame(example_dat)
plot_mech(example_dat$gamma_hat,example_dat$tau_hat,example_dat$sd_u,example_dat$sd_v)
}
}
