% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simest.R
\name{simest}
\alias{simest}
\title{Causal Mediation with SIMEX estimator}
\usage{
simest(gamma_hat, tau_hat, sd_u, b = 1000)
}
\arguments{
\item{gamma_hat}{a vector of the effect of the treatment on the mediator.}

\item{tau_hat}{a vector of the total effect of the treatment on the outcome.}

\item{sd_u}{a vector of the standard error of gamma, the effect of the treatment on the mediator.}

\item{b}{the number of bootstrap replicates. The default number is 1000.}
}
\value{
beta estimated value of beta.

alpha estimated value of alpha.

se_alpha standard error of the estimated alpha.

se_beta standard error of the estimated beta.

p_alpha P-value of alpha.

p_beta P-value of beta.

ave_med the estimated average mediation effects.
}
\description{
Estimation of the beta (key part of the causal mediation effect) with the SIMEX estimator.
}
\examples{
\dontrun{
data(example_dat)  # input data
example_dat <- as.data.frame(example_dat)
simest(example_dat$gamma_hat,example_dat$tau_hat,sd_u) # simex estimator

tmp <- simest(example_dat$gamma_hat,example_dat$tau_hat,example_dat$sd_u) # save the result
tmp$beta  # extract the value
}
}
